requirements.txt
joblib==1.2.0
pandas==1.5.3
pyTelegramBotAPI==4.10.0
python-decouple==3.8
scikit_learn==1.2.2
